+++ src/views/RecIM/index.js
@@ -1,89 +1,22 @@
-import { Grid, Typography, Card, CardHeader, CardContent } from '@material-ui/core/';
-import { useUser } from 'hooks';
-import GordonLoader from 'components/Loader';
-import GordonUnauthorized from 'components/GordonUnauthorized';
-import LeagueListing from './components/LeagueListing';
-import TeamListing from './components/TeamListing';
-import styles from './RecIM.module.css';
-import recimLogo from './recim_logo.jpg';
-
-// CARD - upcoming events
-let upcomingEvents = (
-  <Card>
-    <CardHeader title="Upcoming Rec-IM Events" className={styles.card} />
-    <CardContent>
-      <Typography variant="body1" paragraph>
-        {/* if there are upcoming events, map them here */}
-        <div className={styles.listing}>
-          <LeagueListing leagueId={123456}/>
-        </div>
-        <div className={styles.listing}>
-          <LeagueListing leagueId={12345}/>
-        </div>
-        {/* else "no upcoming events" */}
-        It looks like there aren't any Rec-IM events currently open for registration :(
-      </Typography>
-    </CardContent>
-  </Card>
-);
-
-// CARD - my teams
-let myTeams = (
-  <Card>
-    <CardHeader title="My Teams" className={styles.card} />
-    <CardContent>
-      <Typography variant="body1" paragraph>
-        {/* if I am apart of any active teams, map them here */}
-        <div className={styles.listing}>
-          <TeamListing leagueId={123456} teamId={789}/>
-        </div>
-        <div className={styles.listing}>
-          <TeamListing leagueId={12345} teamId={987}/>
-        </div>
-        {/* else "no teams" */}
-        You're not yet apart of any teams; join one to get started!
-      </Typography>
-    </CardContent>
-  </Card>
-);
+import { Switch, Route } from 'react-router-dom';
+import Home from './views/home';
+import Team from './views/team';
+import League from './views/league';
 
 const RecIM = () => {
-  const { profile, loading } = useUser();
-  // profile hook used for future authentication
-  // Administration privs will use AuthGroups -> example can be found in
-  //           src/components/Header/components/NavButtonsRightCorner
-  if (loading) {
-    return <GordonLoader />;
-  } else if (!profile) {
-    // The user is not logged in
-    return <GordonUnauthorized feature={'the Rec-IM page'} />;
-  } else {
-    return (
-      <>
-        <Grid container alignItems="center">
-          <Grid item>
-            <img src={recimLogo} alt="Rec-IM Logo" width="85em"></img>
-          </Grid>
-          &nbsp;&nbsp;&nbsp;&nbsp;
-          <Grid item>
-            <Typography variant="h5">Rec-IM</Typography>
-            <Typography variant="h6" className={styles.grayText}>
-              <i>"Competition reveals character"</i>
-            </Typography>
-          </Grid>
-        </Grid>
-        <Grid container justifyContent="center" spacing={2}>
-          <Grid item xs={12} md={8}>
-            {upcomingEvents}
-          </Grid>
-          <Grid item xs={12} md={4}>
-            {myTeams}
-          </Grid>
-        </Grid>
-        <Typography variant="subtitle1">Current UserID: {profile.ID}</Typography>
-      </>
-    );
-  }
+  return (
+    <Switch>
+      <Route exact path='/recim'>
+        <Home />
+      </Route>
+      <Route path='/recim/league/:leagueId/team/:teamId'>
+        <Team />
+      </Route>
+      <Route path='/recim/league/:leagueId'>
+        <League />
+      </Route>
+    </Switch>
+  )
 };
 
 export default RecIM;
