+++ src/views/RecIM/views/home/components/LeagueListing/LeagueListing.module.scss
@@ -1,6 +1,6 @@
-@import '../../../../vars';
+@import '../../../../../../vars';
 
-.teamListing {
+.leagueListing {
   background-color: $neutral-light-gray;
   padding: 1em;
   border-radius: 0.5em;
