+++ src/views/RecIM/views/home/components/TeamListing/TeamListing.module.scss
@@ -1,6 +1,6 @@
-@import '../../../../vars';
+@import '../../../../../../vars';
 
-.leagueListing {
+.teamListing {
   background-color: $neutral-light-gray;
   padding: 1em;
   border-radius: 0.5em;
